fix: BrandImage, ProductImage를 Brand, Product에 병합
