--- src/main/java/org/boot/growup/product/service/BrandImageService.java
@@ -1,26 +1,0 @@
-package org.boot.growup.product.service;
-
-import org.boot.growup.product.persist.entity.Brand;
-import org.boot.growup.product.persist.entity.BrandImage;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
-
-@Service
-public interface BrandImageService {
-    /*
-    브랜드 이미지 저장
-     */
-    void postBrandImages(List<MultipartFile> brandImageFiles, Brand brand);
-
-    /*
-    브랜드 이미지 읽어오기
-     */
-    List<BrandImage> getBrandImages(Long id);
-
-    /*
-    브랜드 이미지 수정하기
-     */
-    void patchBrandImages(List<MultipartFile> brandImageFiles, Brand brand);
-}
