+++ src/main/java/org/boot/growup/product/service/ProductService.java
@@ -2,9 +2,12 @@ package org.boot.growup.product.service;
 
 import org.boot.growup.common.constant.AuthorityStatus;
 import org.boot.growup.auth.persist.entity.Customer;
+import org.boot.growup.common.constant.Section;
 import org.boot.growup.product.dto.request.PostProductRequestDTO;
 import org.boot.growup.product.persist.entity.Product;
 import org.boot.growup.auth.persist.entity.Seller;
+import org.boot.growup.product.persist.entity.ProductImage;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -43,4 +46,19 @@ public interface ProductService {
     상품 좋아요 감소
      */
     void deleteProductLike(Long productId, Customer customer);
+
+    /*
+    상품 이미지 등록
+     */
+    void postProductImages(List<MultipartFile> productImageFiles, Product product, Section section);
+
+    /*
+    상품 이미지 읽어오기
+     */
+    List<ProductImage> getProductImages(Long id);
+
+    /*
+    등록된 상품 이미지 수정
+     */
+    void patchProductImages(List<MultipartFile> productImages, Product product, Section section);
 }
