+++ src/test/java/org/boot/growup/source/seller/service/BrandServiceImplTest.java
@@ -3,24 +3,33 @@ package org.boot.growup.source.seller.service;
 import org.boot.growup.common.model.BaseException;
 import org.boot.growup.common.constant.ErrorCode;
 import org.boot.growup.common.constant.AuthorityStatus;
+import org.boot.growup.common.utils.ImageStore;
 import org.boot.growup.product.dto.request.PostBrandRequestDTO;
 import org.boot.growup.product.persist.entity.Brand;
 import org.boot.growup.auth.persist.entity.Seller;
+import org.boot.growup.product.persist.entity.BrandImage;
+import org.boot.growup.product.persist.repository.BrandImageRepository;
 import org.boot.growup.product.persist.repository.BrandRepository;
 import org.boot.growup.auth.persist.repository.SellerRepository;
 import org.boot.growup.product.service.Impl.BrandServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.times;
 
 @SpringBootTest
 @Transactional
@@ -37,6 +46,12 @@ class BrandServiceImplTest {
     @Autowired
     private SellerRepository sellerRepository;
 
+    @Mock
+    private BrandImageRepository brandImageRepository;
+
+    @Mock
+    private ImageStore imageStore;
+
     PostBrandRequestDTO postBrandRequestDTO1;
     PostBrandRequestDTO postBrandRequestDTO2;
     PostBrandRequestDTO postBrandRequestDTO3;
@@ -45,6 +60,11 @@ class BrandServiceImplTest {
 
     @BeforeEach
     void setUp() {
+        File dir = new File("/Users/gnues/Documents/grow/Images/brandImages/");
+        if (!dir.exists()) {
+            dir.mkdirs();
+        }
+
         sellerRepository.deleteAll();  // 기존 데이터 삭제
         brandRepository.deleteAll();
 
@@ -370,5 +390,56 @@ class BrandServiceImplTest {
         //then
         assertNotNull(brandFound);
         assertEquals(brand1, brandFound);
+    }
+
+    @Test
+    public void postBrandImages_Default_Success() {
+        // given
+        Brand brand = new Brand();
+
+        // 두 개의 MockMultipartFile 객체 생성
+        MockMultipartFile file1 = new MockMultipartFile("file", "test1.jpg", "image/jpeg", "test image content 1".getBytes());
+        MockMultipartFile file2 = new MockMultipartFile("file", "test2.jpg", "image/jpeg", "test image content 2".getBytes());
+
+        // Mock된 메소드 동작 설정
+        when(imageStore.createStoreFileName("test1.jpg")).thenReturn("uud-test1.jpg");
+        when(imageStore.createStoreFileName("test2.jpg")).thenReturn("uud-test2.jpg");
+
+        // when
+        brandServiceImpl.postBrandImages(List.of(file1, file2), brand);
+
+        // then
+        verify(imageStore, times(1)).createStoreFileName("test1.jpg");
+        verify(imageStore, times(1)).createStoreFileName("test2.jpg");
+
+        verify(brandImageRepository, times(2)).save(any(BrandImage.class));
+    }
+
+    @Test
+    public void getBrandImages_Default_Success() {
+
+    }
+
+    @Test
+    public void patchBrandImages_Default_Success() {
+        // given
+        Brand brand = Brand.builder().id(1L).build();
+
+        // 두 개의 MockMultipartFile 객체 생성
+        MockMultipartFile file1 = new MockMultipartFile("file", "test1.jpg", "image/jpeg", "test image content 1".getBytes());
+        MockMultipartFile file2 = new MockMultipartFile("file", "test2.jpg", "image/jpeg", "test image content 2".getBytes());
+
+        // Mock된 메소드 동작 설정
+        when(imageStore.createStoreFileName("test1.jpg")).thenReturn("uud-test1.jpg");
+        when(imageStore.createStoreFileName("test2.jpg")).thenReturn("uud-test2.jpg");
+
+        // when
+        brandServiceImpl.patchBrandImages(List.of(file1, file2), brand);
+
+        // then
+        verify(imageStore, times(1)).createStoreFileName("test1.jpg");
+        verify(imageStore, times(1)).createStoreFileName("test2.jpg");
+
+        verify(brandImageRepository, times(2)).save(any(BrandImage.class));
     }
 }
